chore: refactor EVERYTHING.
