+++ packages/sshnpdpy/lib/__init__.py
@@ -1,0 +1,1 @@
+from .sshnpdclient import SSHNPDClient
