--- packages/sshnpdpy/lib/sshnpd.py
@@ -1,339 +1,0 @@
-#!/usr/bin/env python
-from io import StringIO
-import logging
-import os, threading, argparse
-import subprocess
-from queue import Empty, Queue
-from time import sleep
-from uuid import uuid4
-from paramiko import AutoAddPolicy, SSHClient
-from paramiko.ssh_exception import NoValidConnectionsError
-from paramiko.ed25519key import Ed25519Key
-
-
-from socket import socket
-from select import select
-
-from at_client import AtClient
-from at_client.common import AtSign
-from at_client.util import EncryptionUtil
-from at_client.common.keys import AtKey, Metadata, SharedKey
-from at_client.connections.notification.atevents import AtEvent, AtEventType
-
-namespace = "sshnp"
-
-
-def handle_decryption(
-    queue: Queue, client: AtClient, ssh_path, args: argparse.ArgumentParser
-):
-    private_key = ""
-    sshPublicKey = ""
-    ssh_notification_recieved = False
-    while not ssh_notification_recieved or sshPublicKey == "" or private_key == "":
-        try:
-            at_event = queue.get(block=False)
-            event_type = at_event.event_type
-            event_data = at_event.event_data
-
-            # There's defintely a better way to do this
-            # oh well
-            # TODO: fix this
-            if event_type == AtEventType.UPDATE_NOTIFICATION:
-                queue.put(at_event)
-                sleep(1)
-            if event_type != AtEventType.DECRYPTED_UPDATE_NOTIFICATION:
-                continue
-            key = event_data["key"].split(":")[1].split(".")[0]
-            decrypted_value = str(event_data["decryptedValue"])
-            if key == "privatekey":
-                # print(
-                #     f'private key received from ${event_data["from"]} notification id : ${event_data["id"]}')
-                private_key = decrypted_value
-                continue
-
-            if key == "sshpublickey":
-                # print(
-                # f'ssh Public Key received from ${event_data["from"]} notification id : ${event_data["id"]}')
-                sshPublicKey = decrypted_value
-                # // Check to see if the ssh Publickey is already in the file if not append to the ~/.ssh/authorized_keys file
-                writeKey = False
-                with open(f"{ssh_path}/authorized_keys", "r") as read:
-                    filedata = read.read()
-                    if sshPublicKey not in filedata:
-                        writeKey = True
-                with open(f"{ssh_path}/authorized_keys", "w") as write:
-                    if writeKey:
-                        write.write(f"\n{sshPublicKey}")
-                        print("key written")
-                continue
-
-            if key == "sshd":
-                # print(
-                #     f'ssh callback requested from {event_data["from"]} notification id : {event_data["id"]}')
-                ssh_notification_recieved = True
-                callbackArgs = [
-                    at_event,
-                    client,
-                    private_key,
-                    args.manager_atsign,
-                    args.device_atsign,
-                    args.device,
-                    ssh_path,
-                ]
-
-        except Empty:
-            pass
-    return callbackArgs
-
-
-def handle_events(queue: Queue, client: AtClient):
-    while True:
-        try:
-            at_event = queue.get(block=False)
-            event_type = at_event.event_type
-            # Surely this can't be the best way to do this
-            # probably something involving locks because I think both threads are trying to access the queue
-            # the Main thread needs to access the queue AFTER the handle thread has finished with it
-            if event_type == AtEventType.DECRYPTED_UPDATE_NOTIFICATION:
-                queue.put(at_event)
-                sleep(1)
-            else:
-                if event_type == AtEventType.UPDATE_NOTIFICATION:
-                    client.secondary_connection.execute_command(
-                        "notify:remove:" + at_event.event_data["id"]
-                    )
-                client.handle_event(queue, at_event)
-        except Empty:
-            pass
-
-
-def reverse_ssh_exec(ssh_list: list, private_key, sessionID):
-    local_port = ssh_list[0]
-    port = ssh_list[1]
-    username = ssh_list[2]
-    hostname = ssh_list[3]
-    filename = f"/tmp/.{uuid4()}"
-    exitCode = 0
-    if not private_key.endswith("\n"):
-        private_key += "\n"
-    with open(filename, "w") as file:
-        file.write(private_key)
-        subprocess.run(["chmod", "go-rwx", filename])
-    args = [
-        "ssh",
-        f"{username}@{hostname}",
-        "-p",
-        port,
-        "-i",
-        filename,
-        "-R",
-        f"{local_port}:localhost:22",
-        "-t",
-        "-o",
-        "StrictHostKeyChecking=accept-new",
-        "-o",
-        "IdentitiesOnly=yes",
-        "-o",
-        "BatchMode=yes",
-        "-o",
-        "ExitOnForwardFailure=yes",
-        "-v",
-    ]
-    try:
-        process = subprocess.Popen(
-            args, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True
-        )
-        stdout, stderr = process.communicate()
-        exitCode = process.returncode
-        if exitCode != 0:
-            print("ssh session failed for " + username)
-        else:
-            print(
-                "ssh session started for "
-                + username
-                + "@"
-                + hostname
-                + " on port "
-                + port
-            )
-        os.remove(filename)
-    except Exception as e:
-        print(e)
-
-    return exitCode == 0
-
-
-def forward_socket_handler(chan, dest):
-    sock = socket()
-    try:
-        sock.connect(dest)
-    except Exception as e:
-        print(f"Forwarding request to {dest} failed: {e}")
-        return
-
-    print(
-        f"Connected!  Tunnel open {chan.origin_addr} -> {chan.getpeername()} -> {dest}"
-    )
-
-    while True:
-        r, w, x = select([sock, chan], [], [])
-        if sock in r:
-            data = sock.recv(1024)
-            if len(data) == 0:
-                break
-            chan.send(data)
-        if chan in r:
-            data = chan.recv(1024)
-            if len(data) == 0:
-                break
-            sock.send(data)
-    chan.close()
-    sock.close()
-    print(f"Tunnel closed from {chan.origin_addr}")
-
-
-def forward_socket(tp, dest):
-    while True:
-        chan = tp.accept(1000)
-        if chan is None:
-            continue
-
-        threading.Thread(
-            target=forward_socket_handler,
-            args=(chan, dest),
-            daemon=True,
-        ).start()
-
-
-def reverse_ssh_client(ssh_list: list, private_key: str, ssh_path: str):
-    local_port = ssh_list[0]
-    port = ssh_list[1]
-    username = ssh_list[2]
-    hostname = ssh_list[3]
-    print("ssh session started for " + username + " @ " + hostname + " on port " + port)
-    ssh_client = SSHClient()
-    ssh_client.load_system_host_keys(f"{ssh_path}/known_hosts")
-    ssh_client.set_missing_host_key_policy(AutoAddPolicy())
-    file_like = StringIO(private_key)
-    paramiko_log = logging.getLogger("paramiko.transport")
-    paramiko_log.setLevel(logging.DEBUG)
-    paramiko_log.addHandler(logging.StreamHandler())
-    try:
-        pkey = Ed25519Key.from_private_key(file_obj=file_like)
-        ssh_client.connect(
-            hostname=hostname,
-            port=port,
-            username=username,
-            pkey=pkey,
-            allow_agent=False,
-            timeout=10,
-            disabled_algorithms={"pubkeys": ["rsa-sha2-512", "rsa-sha2-256"]},
-        )
-        tp = ssh_client.get_transport()
-        print("Forwarding port " + local_port + " to " + hostname + ":" + port)
-        tp.request_port_forward("", int(local_port))
-        threading.Thread(
-            target=forward_socket,
-            args=(tp, ("localhost", 22)),
-            daemon=True,
-        ).start()
-
-    except NoValidConnectionsError as e:
-        print("Failed to connect: ")
-        for exception in e.errors:
-            print(exception)
-        return False
-    except Exception as e:
-        print(e)
-        return False
-    return True
-
-
-def sshnp_callback(
-    event: AtEvent,
-    at_client: AtClient,
-    private_key,
-    manager_atsign,
-    device_atsign,
-    device,
-    ssh_path,
-):
-    uuid = event.event_data["id"]
-    ssh_list = event.event_data["decryptedValue"].split(" ")
-    iv_nonce = EncryptionUtil.generate_iv_nonce()
-    metadata = Metadata(
-        is_public=False,
-        ttl=10000,
-        ttr=-1,
-        is_encrypted=True,
-        namespace_aware=True,
-        iv_nonce=iv_nonce,
-    )
-    at_key = AtKey(f"{uuid}.", device_atsign)
-    at_key.shared_with = AtSign(manager_atsign)
-    at_key.metadata = metadata
-    at_key.namespace = f"{device}.sshnp"
-    if len(ssh_list) == 5:
-        uuid = ssh_list[4]
-        at_key = AtKey(f"{uuid}", device_atsign)
-        at_key.shared_with = AtSign(manager_atsign)
-        at_key.metadata = metadata
-        at_key.namespace = f".{device}.sshnp"
-
-    ssh_auth = reverse_ssh_client(ssh_list, private_key, ssh_path)
-    if ssh_auth:
-        result = at_client.notify(at_key, "connected")
-        print("sent ssh notification to " + at_key.shared_with.to_string())
-        return True
-    else:
-        print("ssh failed")
-        return False
-
-
-def main():
-    parser = argparse.ArgumentParser("sshnpd")
-    parser.add_argument("--manager", dest="manager_atsign", type=str)
-    parser.add_argument("--atsign", dest="device_atsign", type=str)
-    parser.add_argument("--device", dest="device", type=str)
-    parser.add_argument("--username", dest="username", type=str)
-    args = parser.parse_args()
-    home_dir = ""
-    if os.name == "posix":  # Unix-based systems (Linux, macOS)
-        home_dir = "/root"  # os.path.expanduser("~")
-    elif os.name == "nt":  # Windows
-        home_dir = os.path.expanduser("~")
-    else:
-        raise NotImplementedError("Unsupported operating system")
-
-    commit_log_path = os.path.dirname(
-        f"{home_dir}/.sshnp/{args.device_atsign}/storage/commitLog"
-    )
-    download_path = os.path.dirname(f"{home_dir}/.sshnp/files")
-    ssh_path = f"{home_dir}/.ssh"
-    callbackArgs = []
-    event_queue = Queue(maxsize=20)
-    client = AtClient(AtSign(args.device_atsign), queue=event_queue)
-
-    # Put the username key
-    device_namespace = f"{args.device}.sshnp"
-    username_key = SharedKey(
-        "username", AtSign(args.device_atsign), AtSign(args.manager_atsign)
-    )
-    username_key.namespace = device_namespace
-
-    client.put(username_key, args.username)
-
-    regex = f"{namespace}"
-    threading.Thread(target=client.start_monitor, args=(regex,)).start()
-    threading.Thread(target=handle_events, args=(client.queue, client)).start()
-    callbackArgs = handle_decryption(client.queue, client, ssh_path, args)
-    ssh_connection = sshnp_callback(*callbackArgs)
-    while not ssh_connection:
-        sleep(3)
-        callbackArgs = handle_decryption(client.queue, client, ssh_path, args)
-        ssh_connection = sshnp_callback(*callbackArgs)
-    # TODO: clean up connections and threads lol ( i have lots of faith in the garbage collectors
-
-
-if __name__ == "__main__":
-    main()
